Fix TestBuild.sh to not download, use buildupdate
Developers can use the buildupdate.*.sh scripts to download
dependencies instead of using MSBuild

+++ build/buildupdate.mono.sh
@@ -1,10 +1,25 @@
 #!/bin/bash
 # server=build.palaso.org
 # project=libpalaso
-# build=palaso-precise64-develop-continuous
+# build=palaso-precise64-master-buildupdate
 # root_dir=..
+# $Id: da666a7e5eb1d63b434514279cd14cacd26c730f $
 
 # *** Functions ***
+force=
+
+while getopts f opt; do
+	case $opt in
+	f)
+		force=1
+		;;
+
+	esac
+done
+
+shift $((OPTIND - 1))
+
+
 copy_auto() {
 	where_curl=$(type -P curl)
 	where_wget=$(type -P wget)
@@ -22,7 +37,7 @@ copy_auto() {
 
 copy_curl() {
 	echo "curl: $2 <= $1"
-	if [ -e "$2" ]
+	if [ -e "$2" ] && [ "$force" != "1" ]
 	then
 		curl -# -L -z $2 -o $2 $1
 	else
@@ -40,10 +55,10 @@ copy_wget() {
 }
 
 # *** Results ***
-# build: palaso-precise64-develop-continuous (bt334)
+# build: palaso-precise64-master-buildupdate (bt345)
 # project: libpalaso
-# URL: http://build.palaso.org/viewType.html?buildTypeId=bt334
-# VCS: https://github.com/sillsdev/libpalaso.git [develop]
+# URL: http://build.palaso.org/viewType.html?buildTypeId=bt345
+# VCS: https://github.com/sillsdev/libpalaso.git [master]
 # dependencies:
 # [0] build: L10NSharp Mono continuous (bt271)
 #     project: L10NSharp
@@ -64,14 +79,14 @@ copy_wget() {
 #     URL: http://build.palaso.org/viewType.html?buildTypeId=bt281
 #     clean: false
 #     revision: latest.lastSuccessful
-#     paths: {"*.dll"=>"lib\\DebugMono", "*.config"=>"lib\\DebugMono"}
+#     paths: {"icu.net.*"=>"lib\\DebugMono"}
 #     VCS: https://github.com/sillsdev/icu-dotnet [master]
 # [3] build: icucil-precise64-Continuous (bt281)
 #     project: Libraries
 #     URL: http://build.palaso.org/viewType.html?buildTypeId=bt281
 #     clean: false
 #     revision: latest.lastSuccessful
-#     paths: {"*.dll"=>"lib\\ReleaseMono", "*.config"=>"lib\\ReleaseMono"}
+#     paths: {"icu.net.*"=>"lib\\ReleaseMono"}
 #     VCS: https://github.com/sillsdev/icu-dotnet [master]
 
 # make sure output directories exist
@@ -85,6 +100,8 @@ copy_auto http://build.palaso.org/guestAuth/repository/download/bt271/latest.las
 copy_auto http://build.palaso.org/guestAuth/repository/download/bt271/latest.lastSuccessful/L10NSharp.dll.mdb ../lib/DebugMono/L10NSharp.dll.mdb
 copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll ../lib/DebugMono/icu.net.dll
 copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll.config ../lib/DebugMono/icu.net.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll.mdb ../lib/DebugMono/icu.net.dll.mdb
 copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll ../lib/ReleaseMono/icu.net.dll
 copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll.config ../lib/ReleaseMono/icu.net.dll.config
+copy_auto http://build.palaso.org/guestAuth/repository/download/bt281/latest.lastSuccessful/icu.net.dll.mdb ../lib/ReleaseMono/icu.net.dll.mdb
 # End of script
